+++ imagery/i.segment/flag.c
@@ -20,7 +20,7 @@ FLAG *flag_create(int nrows, int ncols)
     if (!new_flag->array)
         G_fatal_error(_("Out of memory!"));
 
-    temp = (unsigned char *)G_malloc(nrows * new_flag->leng *
+    temp = (unsigned char *)G_malloc((size_t)nrows * new_flag->leng *
                                      sizeof(unsigned char));
 
     if (!temp)
